improve calculations by using floor div
The Problem is, that floats are not arbitrary, but have a fixed sized. This leads to issues with big q and p values. To fix this, floor division is used, as its outputs integers and the maximum (theoretical) 1 difference between the floor result and the real one, has not such big impacts by huge values